Add logic for sending websocket message and added websocket config for the messages
